test: address characterization review feedback
